% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_functions.R
\name{lp_irf}
\alias{lp_irf}
\title{Local Projection VAR impulse responses}
\usage{
lp_irf(y, h, max.lags)
}
\arguments{
\item{y}{\code{xts} object with the data}

\item{h}{the impulse response (IRF) horizon}

\item{max.lags}{the maximum number of lags to be considered in the
VAR}
}
\description{
VAR impulse responses using the local projection method of Jorda
(2005). The lags within each horizon will be chosen by the AIC
}

