% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us.macro.data}
\alias{us.macro.data}
\title{US macro data (xts object)}
\format{An xts object with monthly US macro data
\describe{
    \item{INDRPO}{log first difference of industrial production}
    \item{CPIFESL}{log first difference of CPI less food and energy}
    \item{UNRATE}{the unemployment rate}
    \item{FEDFUNDS}{the fed funds rate}
}}
\usage{
us.macro.data
}
\description{
A us macro data dataset as an xts object from 1957-02-01 to
2008-01-01. Data were downloaded from the FRED. The variable names
correspond to FRED symbols
}
\keyword{datasets}

