% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_functions.R
\name{var_aic}
\alias{var_aic}
\title{AIC from a VAR regression}
\usage{
var_aic(yy, xx)
}
\arguments{
\item{yy}{the left-hand-side variables}

\item{xx}{the right hand side variables}
}
\value{
The AIC for the VAR
}
\description{
\code{yy} and \code{xx} have to have the same dimensions and all
lags must already be created in \code{xx}
}

